Updating PlotActivity, FormularioActivity, and FolderActivity
The Plot Activity, Form Activity, and Folder Activity files have been updated for better viewing and new tools for better navigation. The save button has also been correctly implemented to link to the folders of the created patients, in order to implement the temperature vs. time graph in the future.